Moved makeFIle function to ControllerMain
Added main.fxml and ControllerMain + moved makeFIle function from Main to ControllerMain
