+++ src/com/Hammak/main.fxml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.AnchorPane?>
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="com.Hammak.ControllerMain"
+            prefHeight="400.0" prefWidth="600.0">
+
+</AnchorPane>
